requirements.txt
Mastodon.py==1.4.0
beautifulsoup4==4.6.0
